feat: Correccion Pull Request 7
