+++ lib/features/login/domain/datasources/auth_data_source.dart
@@ -1,0 +1,8 @@
+import 'package:firebase_auth/firebase_auth.dart';
+
+abstract class AuthDataSource {
+  Future<User?> login(String email, String password);
+  Future<User?> register(String email, String password);
+  Future<void> logout();
+  Stream<User?> authStateChanges();
+}
